final mult school

+++ database/migrations/2026_11_10_211111_create_subject_class_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('subject_class')) return;
+
         Schema::create('subject_class', function (Blueprint $table) {
         $table->id();
         $table->foreignId('subject_id')->constrained()->onDelete('cascade');
